Fix #2 (Storing the current teaser in the language file is nonsense)
Storing this information in the language file is indeed nonsense.
Besides the fact that plugins are not supposed to modify the language
files, constantly updating the language files will cause OPcache out-
of-memory conditions causing a restart, which is bad.

Instead, we store that temporary information in the session.  That
means that it won't prevail over a long period of time (i.e. the next
admin session), but that might actually be more sensible behavior.